requirements.txt
Pillow==8.3.1
numpy<=1.19
tensorflow==2.3
matplotlib==3.4.2
tqdm==4.61.2
opencv-python==4.5.2.54
scikit-image== 0.18.2
setuptools~=57.1.0
scipy==1.4.1
scikit-learn~=0.24.2
tf-slim==1.1.0